split the db_system test.
fmt code.

refactoring to avoid using testsuite.

remove unneeded comments.
